Extend bank to multiple accounts
Note: my initial implementation actually had a bug where I was
specifying the offsets in `Acct` in bytes instead of bits (`n*8` instead
of `n*64`), which I eventually caught when I couldn't prove Acct(n)
was in the domain.

In the meantime, with this bug and without an invariant, verification
was extremely painful (the file would take 30-45s to compile, and VSCode
was so slow in getting results it always had errors in the wrong
places).