Add inode decoding benchmarks
Dafny                  1000 ns/op
Manual decode          234 ns/op
Decode to native types 48 ns/op

The manual decode implementation has the same API as Dafny, but uses
marshal and uses native types internally (only converting to a dafny.Seq
at the end). If we change the representation of inodes to be a regular
struct with a native []uint64, then decoding is just 48 ns/op.