Change how unused data is represented
At the block level an inode is now always a seq<Block> that's exactly 15
blocks long; to avoid allocating blocks to represent that data, we allow
bn=0 to represent a zero block. This is good preparation for indirect
inodes and actually results in a simpler representation (there are no
longer two lengths involved for an inode, only holes that have to be
allocated to write any data to them).