Use uint64 in Pos
This affects a lot of code -it effectively maintains an invariant within
a Pos that's independent of the config. However Pos is used at runtime
to compute inode and indirect block offsets so it's a big win to avoid
all of this use of big nums. It's really annoying to do this "properly"
with separate specification and runtime types and relations between
them, so we just fix the spec type to build in this stronger invariant.