Allow Dirents to be more than one block
We now enforce the restriction in MemDirents and in dir_fs, in
preparation for large directories. Dirents has a bunch of useful theory
for pure DirEnt sequences that's well-organized without factoring out
properties of seq<DirEnt> (actually there's a bunch of theory in the
file for those, too, but right now they just serve as lemmas).