Fix a big issue with readDirents
We no longer read the entire directory and then close it; we instead
need to leave it open for future directory operations, and close it only
after reading or writing to it.

One unfortunate side effect of this requirement is that we cannot have a
directory and file open at the same time; most of the time this is fine,
but it must be managed correctly by every operation.