Fix write hole hack to truly be atomic
Wow this was subtle and I didn't think hard enough about it - there was
a time-of-check to time-of-use bug where if a write failed because it
required a hole, we would try again and set the size. However, in the
meantime the size might have grown, and the hack would then truncate the
size of the file. The only proper way to implement the hack is to
atomically check the size of the file and grow it in the same
transaction, so we now move that code into Dafny.

Writing any large file on Linux actually triggers this bug because many
concurrent writes attempt to fill holes.