Implement size hints to reduce cost of freeing
The idea is to track (in an unverified way) where space is likely to be
used and only free that space. There's some signalling so that when
space is needed, that space is cleared. The biggest danger is that if
something is wrong the space will never be cleared and the client might
keep getting ERRJUKEBOX.